separating the spec tests as per the comments
